split of validation_globals
