+++ src/Makefile.am
@@ -186,6 +186,8 @@ BITCOIN_CORE_H = \
   utiltime.h \
   validation.h \
   validation_globals.h \
+  validation_block_utils.h \
+  validation_disk_utils.h \
   validationinterface.h \
   versionbits.h \
   walletinitinterface.h \
@@ -257,7 +259,8 @@ libbitcoin_server_a_SOURCES = \
   txmempool.cpp \
   ui_interface.cpp \
   validation.cpp \
-  validation_globals.cpp \
+  validation_block_utils.cpp \
+  validation_disk_utils.cpp \
   validationinterface.cpp \
   versionbits.cpp \
   $(BITCOIN_CORE_H)
