+++ src/index/txindex.cpp
@@ -7,7 +7,7 @@
 #include <ui_interface.h>
 #include <util.h>
 #include <validation_globals.h>
-
+#include <validation_disk_utils.h>
 #include <boost/thread.hpp>
 
 constexpr char DB_BEST_BLOCK = 'B';
