+++ src/validation_disk_utils.h
@@ -1,0 +1,27 @@
+// Copyright (c) 2018 The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BITCOIN_VALIDATION_DISK_UTILS_H
+#define BITCOIN_VALIDATION_DISK_UTILS_H
+
+#include <validation_globals.h>
+#include <primitives/block.h>
+#include <protocol.h>
+#include <stdio.h>
+
+
+FILE* OpenDiskFile(const CDiskBlockPos &pos, const char *prefix, bool fReadOnly);
+
+/** Open a block file (blk?????.dat) */
+FILE* OpenBlockFile(const CDiskBlockPos &pos, bool fReadOnly = false);
+
+/** Functions for disk access for blocks */
+bool ReadBlockFromDisk(CBlock& block, const CDiskBlockPos& pos, const Consensus::Params& consensusParams);
+bool ReadBlockFromDisk(CBlock& block, const CBlockIndex* pindex, const Consensus::Params& consensusParams);
+bool ReadRawBlockFromDisk(std::vector<uint8_t>& block, const CDiskBlockPos& pos, const CMessageHeader::MessageStartChars& message_start);
+bool ReadRawBlockFromDisk(std::vector<uint8_t>& block, const CBlockIndex* pindex, const CMessageHeader::MessageStartChars& message_start);
+
+fs::path GetBlockPosFilename(const CDiskBlockPos &pos, const char *prefix);
+
+#endif // BITCOIN_VALIDATION_DISK_UTILS_H
